all tags are working, some css

+++ db/migrate/20180716160450_create_prices.rb
@@ -2,7 +2,7 @@ class CreatePrices < ActiveRecord::Migration[5.2]
   def change
     create_table :prices do |t|
       t.integer :nombre
-      t.integer :unite
+      t.float :unite
       t.references :fairepart, foreign_key: true
 
       t.timestamps
